+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 
 ## (Unreleased) rocSOLVER
 ### Added
+- Partial eigenvalue decomposition routines:
+    - STEBZ
 - Package generation for test and benchmark executables on all supported OSes using CPack.
 - Added tests for multi-level logging
 - File/Folder Reorg